More debug cleanliness and some fixes, though more bugs have crept up in int handling.

+++ drafts/libft/src/ft_digits.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-int		ft_digits(int n)
+int		ft_digits(intmax_t n)
 {
 	return (ft_digits_base(n, 10));
 }
